requirements.txt
Flask==2.0.2

pillow~=8.4.0
numpy~=1.21.4